updated import in test due to renaming
