+++ packages/material/test/renderers/integer.field.test.tsx
@@ -11,7 +11,7 @@ import {
   update
 } from '@jsonforms/core';
 import IntegerField, { integerFieldTester } from '../../src/fields/material-integer.field';
-import HorizontalLayoutRenderer from '../../src/layouts/material-horizontal.layout';
+import HorizontalLayoutRenderer from '../../src/layouts/MaterialHorizontalLayout';
 import {
   change,
   findRenderedDOMElementWithTag,
