+++ packages/material/test/renderers/text.field.test.tsx
@@ -11,7 +11,7 @@ import {
   update
 } from '@jsonforms/core';
 import TextField, { textFieldTester, } from '../../src/fields/material-text.field';
-import HorizontalLayoutRenderer from '../../src/layouts/material-horizontal.layout';
+import HorizontalLayoutRenderer from '../../src/layouts/MaterialHorizontalLayout';
 import {
   change,
   findRenderedDOMElementWithTag,
